+++ Kamienica/src/main/java/kamienica/feature/invoice/InvoiceWaterDAOImpl.java
@@ -2,17 +2,16 @@ package kamienica.feature.invoice;
 
 import java.util.List;
 
-import javax.transaction.Transactional;
-
 import org.hibernate.Query;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import kamienica.dao.AbstractDao;
 import kamienica.feature.payment.PaymentStatus;
 
 @Repository("invoiceWater")
 @Transactional
-public class InvoiceWaterDAOImpl extends AbstractDao<Integer, InvoiceWater> implements InvoiceDao<InvoiceWater> {
+public class InvoiceWaterDAOImpl extends AbstractDao<Long, InvoiceWater> implements InvoiceDao<InvoiceWater> {
 
 	@Override
 	public InvoiceWater getLatest() {
